Improve generated code and add comments

+++ modules/swagger-codegen/src/main/java/io/swagger/codegen/languages/LoopBackSwiftClientCodegen.java
@@ -7,7 +7,6 @@ import io.swagger.models.Swagger;
 import io.swagger.models.properties.*;
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.lang.WordUtils;
-import org.joda.time.DateTime;
 
 import java.io.File;
 import java.util.*;
@@ -20,7 +19,7 @@ public class LoopBackSwiftClientCodegen extends DefaultCodegen implements Codege
 
     private static final String REALM_PRIMARY_KEY = "realmPrimaryKey";
 
-    private static final String GENERATE_TOP_LEVEL_API = "generateTopLevelAPI";
+    private static final String GENERATE_DATA_API = "generateDataAPI";
 
     private static final String PROJECT_NAME = "projectName";
 
@@ -42,7 +41,7 @@ public class LoopBackSwiftClientCodegen extends DefaultCodegen implements Codege
 
     private boolean useRealm = false;
 
-    private boolean generateTopLevelAPI = false;
+    private boolean generateDataAPI = false;
 
     @Override
     public CodegenType getTag() {
@@ -173,16 +172,15 @@ public class LoopBackSwiftClientCodegen extends DefaultCodegen implements Codege
             additionalProperties.put(PROJECT_LICENSE, "The MIT License (MIT)");
         }
 
-        useRealm = (Boolean) additionalProperties.get(USE_REALM);
-        if (useRealm) {
-            if (!additionalProperties.containsKey(GENERATE_TOP_LEVEL_API)) {
-                additionalProperties.put(GENERATE_TOP_LEVEL_API, false);
-            } else {
-                additionalProperties.put(GENERATE_TOP_LEVEL_API, additionalProperties.get(GENERATE_TOP_LEVEL_API).equals("true"));
-            }
-
-            generateTopLevelAPI = (Boolean) additionalProperties.get(USE_REALM);
+        if (!additionalProperties.containsKey(GENERATE_DATA_API)) {
+            additionalProperties.put(GENERATE_DATA_API, false);
+        } else {
+            additionalProperties.put(GENERATE_DATA_API, additionalProperties.get(GENERATE_DATA_API).equals("true"));
         }
+
+        useRealm = (Boolean) additionalProperties.get(USE_REALM);
+
+        generateDataAPI = (Boolean) additionalProperties.get(GENERATE_DATA_API);
 
         modelId = additionalProperties.get(REALM_PRIMARY_KEY).toString();
 
@@ -202,7 +200,7 @@ public class LoopBackSwiftClientCodegen extends DefaultCodegen implements Codege
             typeMapping.put("object", "NSData");
         }
 
-        if (generateTopLevelAPI) {
+        if (generateDataAPI) {
             topLevelTemplateFiles.put("DataAPI.mustache", ".swift");
 
             supportingFiles.add(new SupportingFile("UnsupportedOperation.mustache", utilFileFolder(), "UnsupportedOperation.swift"));
@@ -432,19 +430,21 @@ public class LoopBackSwiftClientCodegen extends DefaultCodegen implements Codege
             return;
         }
 
-        String resourcePath = "/" + operations.get(0).path.split("/")[1];
+        String resourcePath = generateDataAPI ? "/" + operations.get(0).path.split("/")[1] : null;
         String countReturnType = "Int";
         String deleteReturnType = "Bool";
 
         for (CodegenOperation operation: operations) {
-            if (operation.path.equals(resourcePath)) {
-                operation.path = null;
+            if (resourcePath != null) {
+                if (operation.path.equals(resourcePath)) {
+                    operation.path = null;
 
-            } else {
-                String[] pathParts = operation.path.split(resourcePath);
+                } else {
+                    String[] pathParts = operation.path.split(resourcePath);
 
-                if (pathParts.length > 1) {
-                    operation.path = pathParts[1].substring(1);
+                    if (pathParts.length > 1) {
+                        operation.path = pathParts[1].substring(1);
+                    }
                 }
             }
 
@@ -454,34 +454,36 @@ public class LoopBackSwiftClientCodegen extends DefaultCodegen implements Codege
                 operation.successfulResponse.isListContainer = true;
             }
 
-            if (operation.operationId.equals("create")) {
-                api.put("hasCreate", true);
-            }
+            if (generateDataAPI) {
+                if (operation.operationId.equals("create")) {
+                    api.put("hasCreate", true);
+                }
 
-            if (operation.operationId.equals("upsert")) {
-                api.put("hasUpsert", true);
-            }
+                if (operation.operationId.equals("upsert")) {
+                    api.put("hasUpsert", true);
+                }
 
-            if (operation.operationId.equals("findById")) {
-                api.put("hasFindById", true);
-            }
+                if (operation.operationId.equals("findById")) {
+                    api.put("hasFindById", true);
+                }
 
-            if (operation.operationId.equals("findOne")) {
-                api.put("hasFindOne", true);
-            }
+                if (operation.operationId.equals("findOne")) {
+                    api.put("hasFindOne", true);
+                }
 
-            if (operation.operationId.equals("find")) {
-                api.put("hasFind", true);
-            }
+                if (operation.operationId.equals("find")) {
+                    api.put("hasFind", true);
+                }
 
-            if (operation.operationId.equals("count")) {
-                api.put("hasCount", true);
-                countReturnType = operation.successfulResponse.dataType;
-            }
+                if (operation.operationId.equals("count")) {
+                    api.put("hasCount", true);
+                    countReturnType = operation.successfulResponse.dataType;
+                }
 
-            if (operation.operationId.equals("deleteById")) {
-                api.put("hasDeleteById", true);
-                deleteReturnType = operation.successfulResponse.dataType;
+                if (operation.operationId.equals("deleteById")) {
+                    api.put("hasDeleteById", true);
+                    deleteReturnType = operation.successfulResponse.dataType;
+                }
             }
         }
